fixes #1390 - use UNSIGNED INT for ip; since the upper 32 bits is unused, the ALTER statement simply truncates to the lower 32 bits

+++ core/Piwik.php
@@ -716,7 +716,7 @@ class Piwik
 									  parameter_names_default_values TEXT NULL,
 									  parameter_values TEXT NULL,
 									  execution_time FLOAT NULL,
-									  caller_ip BIGINT UNSIGNED NULL,
+									  caller_ip INT UNSIGNED NULL,
 									  timestamp TIMESTAMP NULL,
 									  returned_value TEXT NULL,
 									  PRIMARY KEY(idlogger_api_call)
@@ -791,7 +791,7 @@ class Piwik
 							  config_gears TINYINT(1) NOT NULL,
 							  config_silverlight TINYINT(1) NOT NULL,
 							  config_cookie TINYINT(1) NOT NULL,
-							  location_ip BIGINT UNSIGNED NOT NULL,
+							  location_ip INT UNSIGNED NOT NULL,
 							  location_browser_lang VARCHAR(20) NOT NULL,
 							  location_country CHAR(3) NOT NULL,
 							  location_continent CHAR(3) NOT NULL,
